refactor: move fee input to review step - validator registration
